requirements.txt
flake8==7.0.0
flake8-docstrings==1.5.0
jupyter==1.0.0
python-dotenv==1.0.1
pandas==2.2.1
langchain
langchain-openai
msgraph-sdk
streamlit
fastapi
uvicorn
langchain-community